fixes int32 issue

+++ corl/wc_data/input_file2.py
@@ -182,7 +182,7 @@ def _load_infer_data(task):
     data = _loadData(path, name)
     # features, seqlens, code, klid, refs, idx
     return np.array(data['features'], 'f'), np.array(data['seqlens'], 'i'), data['code'], \
-        data['klid'], np.array(data['refs'], 'U'), int(task[1])
+        np.int32(data['klid']), np.array(data['refs'], 'U'), np.int32(task[1])
 
 
 def _read_meta_config(file_dir):
